+++ src/aws/constants.ts
@@ -1,0 +1,59 @@
+// AWS regions
+export const AWS_REGION_EAST1 = 'us-east-1'; // Virginia, US
+export const AWS_REGION_EAST2 = 'us-east-2'; // Ohio, US
+export const AWS_REGION_WEST1 = 'us-west-1'; // Northern California, US
+export const AWS_REGION_WEST2 = 'us-west-2'; // Oregon, US
+
+export const AWS_REGION_EU_CENTRAL1 = 'eu-central-1'; // Frankfurt, DE
+export const AWS_REGION_EU_WEST1 = 'eu-west-1'; // Ireland, IE
+
+export const AWS_REGION_AP_SE1 = 'ap-southeast-1'; // Singapore, SG
+export const AWS_REGION_AP_SE2 = 'ap-southeast-2'; // Sydney, AU
+export const AWS_REGION_AP_SE4 = 'ap-southeast-4'; // Melbourne, AU
+export const AWS_REGION_AP_NORTHEAST1 = 'ap-northeast-1'; // Tokyo, JP
+export const AWS_REGION_AP_SOUTH1 = 'ap-south-1'; // Mumbai, IN
+
+export const AWS_REGION_SA_EAST1 = 'sa-east-1'; // São Paulo, BR
+
+export const AWS_REGION_AF_SOUTH1 = 'af-south-1'; // Cape Town, ZA
+
+export const AWS_REGIONS = {
+  AWS_REGION_EAST1,
+  AWS_REGION_EAST2,
+  AWS_REGION_WEST1,
+  AWS_REGION_WEST2,
+  AWS_REGION_EU_CENTRAL1,
+  AWS_REGION_EU_WEST1,
+  AWS_REGION_AP_SE1,
+  AWS_REGION_AP_SE2,
+  AWS_REGION_AP_SE4,
+  AWS_REGION_AP_NORTHEAST1,
+  AWS_REGION_AP_SOUTH1,
+} as const;
+
+export type AWSRegionKey = keyof typeof AWS_REGIONS;
+export type AWSRegion = (typeof AWS_REGIONS)[AWSRegionKey] | '';
+
+export const DEFAULT_REGION_US: AWSRegion = AWS_REGION_EAST2;
+export const DEFAULT_REGION_EAST: AWSRegion = AWS_REGION_EAST2;
+export const DEFAULT_REGION_WEST: AWSRegion = AWS_REGION_WEST2;
+
+export const DEFAULT_REGION_EU: AWSRegion = AWS_REGION_EU_CENTRAL1;
+export const DEFAULT_REGION_EU_CENTRAL: AWSRegion = AWS_REGION_EU_CENTRAL1;
+export const DEFAULT_REGION_EU_WEST: AWSRegion = AWS_REGION_EU_WEST1;
+
+export const DEFAULT_REGION_AP: AWSRegion = AWS_REGION_AP_SE1;
+export const DEFAULT_REGION_AP_SE: AWSRegion = AWS_REGION_AP_SE1;
+export const DEFAULT_REGION_AP_NORTHEAST: AWSRegion = AWS_REGION_AP_NORTHEAST1;
+export const DEFAULT_REGION_AP_SOUTH: AWSRegion = AWS_REGION_AP_SOUTH1;
+
+// AWS accounts
+export const DEFAULT_ACCOUNT_US_NONPROD = '122703641091';
+export const DEFAULT_ACCOUNT_US_PROD = '123456789012';
+
+export const DEFAULT_ACCOUNT_NONPROD = DEFAULT_ACCOUNT_US_NONPROD; // global default
+export const DEFAULT_ACCOUNT_PROD = DEFAULT_ACCOUNT_US_PROD; // global default
+
+// AWS ARNs
+
+// AWS IAM/Roles